+++ satellite/gracefulexit/endpoint_test.go
@@ -396,66 +396,6 @@ func TestFailureUplinkSignature(t *testing.T) {
 	})
 }
 
-func TestSignExitCompleted(t *testing.T) {
-	testplanet.Run(t, testplanet.Config{
-		SatelliteCount:   1,
-		StorageNodeCount: 1,
-		UplinkCount:      1,
-	}, func(t *testing.T, ctx *testcontext.Context, planet *testplanet.Planet) {
-		satellite := planet.Satellites[0]
-		node := planet.StorageNodes[0]
-		finishedAt := time.Now().UTC()
-		signer := signing.SignerFromFullIdentity(satellite.Identity)
-		signee := signing.SigneeFromPeerIdentity(satellite.Identity.PeerIdentity())
-
-		unsigned := &pb.ExitCompleted{
-			SatelliteId: satellite.ID(),
-			NodeId:      node.ID(),
-			Completed:   finishedAt,
-		}
-		signed, err := signing.SignExitCompleted(ctx, signer, unsigned)
-		require.NoError(t, err)
-
-		err = signing.VerifyExitCompleted(ctx, signee, signed)
-		require.NoError(t, err)
-
-		signed.SatelliteId = testrand.NodeID()
-
-		err = signing.VerifyExitCompleted(ctx, signee, signed)
-		require.Error(t, err)
-	})
-}
-
-func TestSignExitFailed(t *testing.T) {
-	testplanet.Run(t, testplanet.Config{
-		SatelliteCount:   1,
-		StorageNodeCount: 1,
-		UplinkCount:      1,
-	}, func(t *testing.T, ctx *testcontext.Context, planet *testplanet.Planet) {
-		satellite := planet.Satellites[0]
-		node := planet.StorageNodes[0]
-		finishedAt := time.Now().UTC()
-		signer := signing.SignerFromFullIdentity(satellite.Identity)
-		signee := signing.SigneeFromPeerIdentity(satellite.Identity.PeerIdentity())
-
-		unsigned := &pb.ExitFailed{
-			SatelliteId: satellite.ID(),
-			NodeId:      node.ID(),
-			Failed:      finishedAt,
-			Reason:      pb.ExitFailed_INACTIVE_TIMEFRAME_EXCEEDED,
-		}
-		signed, err := signing.SignExitFailed(ctx, signer, unsigned)
-		require.NoError(t, err)
-
-		err = signing.VerifyExitFailed(ctx, signee, signed)
-		require.NoError(t, err)
-
-		signed.Reason = pb.ExitFailed_OVERALL_FAILURE_PERCENTAGE_EXCEEDED
-
-		err = signing.VerifyExitFailed(ctx, signee, signed)
-		require.Error(t, err)
-	})
-}
 func TestSuccessPointerUpdate(t *testing.T) {
 	testTransfers(t, 1, func(ctx *testcontext.Context, nodeFullIDs map[storj.NodeID]*identity.FullIdentity, satellite *testplanet.SatelliteSystem, processClient exitProcessClient, exitingNode *storagenode.Peer, numPieces int) {
 		var recNodeID storj.NodeID
